==> 2-descriptive-outcomes/DHS_figures/0-run-descriptive-outcomes-DHS.R <==

rm(list=ls())
source(paste0(here::here(), "/0-config.R"))




source(here("2-descriptive-outcomes/DHS_figures/01_DHS_plot_dhsdata_mean_stunting.R"))
source(here("2-descriptive-outcomes/DHS_figures/02_DHS_plot_wasting_mean_estimation.R"))
source(here("2-descriptive-outcomes/DHS_figures/03_DHS_plot_wasting_spline_estimation.R"))
source(here("2-descriptive-outcomes/DHS_figures/03_DHSplot_kidata_density.R"))
source(here("2-descriptive-outcomes/DHS_figures/04_DHSplot_dhsdata_density.R"))
source(here("2-descriptive-outcomes/DHS_figures/05_DHSplot_density_medians.R"))
